services/orchestra-parser/pyproject.toml
[tool.poetry]
name = "orchestra_parser"
version = "0.1.0"
description = ""
authors = ["Fabrizio Armango <fabrizio.armango@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
lark-parser = "^0.11.2"
rdflib = "^5.0.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[tool.poetry.scripts]
gen-grammar = ""
gen-sample = "orchestra_parser.tests.gen_sample_orcx:start"
orcx2json = "orchestra_parser.orcx2json:start"
json2rdf = "orchestra_parser.json2rdf:start"

[tool.poe.tasks]
test1  = "poetry run orcx2json tests/"